fix(suite-native): passphrase mismatch on trezor screen
