+++ suite-native/device-authorization/src/index.ts
@@ -3,4 +3,5 @@ export * from './utils';
 export * from './hooks/useAuthorizationGoBack';
 export * from './hooks/useHandleDeviceRequestsPassphrase';
 export * from './hooks/useHandleDuplicatePassphrase';
+export * from './hooks/useHandlePassphraseMismatch';
 export * from './passphraseThunks';
