+++ suite-native/module-authorize-device/src/screens/passphrase/PassphraseConfirmOnTrezorScreen.tsx
@@ -17,13 +17,7 @@ import {
     selectIsDeviceConnectedAndAuthorized,
     selectIsDeviceDiscoveryActive,
 } from '@suite-common/wallet-core';
-import {
-    cancelPassphraseAndSelectStandardDeviceThunk,
-    retryPassphraseAuthenticationThunk,
-    selectHasPassphraseMismatchError,
-} from '@suite-native/device-authorization';
-import { useAlert } from '@suite-native/alerts';
-import { EventType, analytics } from '@suite-native/analytics';
+import { useHandlePassphraseMismatch } from '@suite-native/device-authorization';
 
 import { DeviceT3T1Svg } from '../../assets/passphrase/DeviceT3T1Svg';
 import { PassphraseScreenWrapper } from '../../components/passphrase/PassphraseScreenWrapper';
@@ -43,13 +37,12 @@ export const PassphraseConfirmOnTrezorScreen = () => {
     const isDeviceConnectedAndAuthorized = useSelector(selectIsDeviceConnectedAndAuthorized);
     const isDiscoveryActive = useSelector(selectIsDeviceDiscoveryActive);
     const device = useSelector(selectDevice);
-    const hasPassphraseMismatchError = useSelector(selectHasPassphraseMismatchError);
-
-    const { showAlert } = useAlert();
 
     // If this screen was present during authorizing device with passphrase for some feature,
     // on success, this hook will close the stack and go back
     useRedirectOnPassphraseCompletion();
+
+    useHandlePassphraseMismatch();
 
     useEffect(() => {
         if (isDeviceConnectedAndAuthorized && isDiscoveryActive) {
@@ -62,42 +55,6 @@ export const PassphraseConfirmOnTrezorScreen = () => {
             );
         }
     }, [device, dispatch, isDeviceConnectedAndAuthorized, isDiscoveryActive, navigation]);
-
-    useEffect(() => {
-        // Wrong passphrase was entered during verifying empty wallet
-        if (hasPassphraseMismatchError) {
-            analytics.report({ type: EventType.PassphraseMismatch });
-            showAlert({
-                title: (
-                    <Translation id="modulePassphrase.emptyPassphraseWallet.verifyEmptyWallet.passphraseMismatchAlert.title" />
-                ),
-                description: (
-                    <Translation id="modulePassphrase.emptyPassphraseWallet.verifyEmptyWallet.passphraseMismatchAlert.description" />
-                ),
-                primaryButtonTitle: (
-                    <Translation id="modulePassphrase.emptyPassphraseWallet.verifyEmptyWallet.passphraseMismatchAlert.primaryButton" />
-                ),
-                onPressPrimaryButton: () => {
-                    navigation.navigate(AuthorizeDeviceStackRoutes.PassphraseForm);
-                    dispatch(retryPassphraseAuthenticationThunk());
-                },
-                primaryButtonVariant: 'redBold',
-                secondaryButtonTitle: (
-                    <Translation id="modulePassphrase.emptyPassphraseWallet.verifyEmptyWallet.passphraseMismatchAlert.secondaryButton" />
-                ),
-                onPressSecondaryButton: () => {
-                    dispatch(cancelPassphraseAndSelectStandardDeviceThunk());
-                    analytics.report({
-                        type: EventType.PassphraseExit,
-                        payload: { screen: AuthorizeDeviceStackRoutes.PassphraseConfirmOnTrezor },
-                    });
-                },
-                secondaryButtonVariant: 'redElevation0',
-                icon: 'warningTriangleLight',
-                pictogramVariant: 'red',
-            });
-        }
-    }, [dispatch, hasPassphraseMismatchError, navigation, showAlert]);
 
     return (
         <PassphraseScreenWrapper>
